Added setNeedsSave for XCSnippet
Now instead of trying to save immediately anytime a property is updated, setNeedsSave is called instead, which waits 1 second and only is called once even if subsequent calls happen in that one second. The benefit of this is that if multiple properties are being set at a time, the snippet will only be saved once, which solves several race condition bugs.